Implemented parsing for variable references and integer literals
